Use double quotes for JS files to avoid confused Go developers

+++ frontend/.eslintrc.js
@@ -14,7 +14,7 @@ module.exports = {
         'comma-dangle': ['error', 'always-multiline'],
         indent: ['error', 4],
         'linebreak-style': ['error', 'unix'],
-        quotes: ['error', 'single'],
+        quotes: ['error', 'double'], // Easier for Go developers!
         semi: ['error', 'always'],
         'no-unused-vars': ['warn'],
         'no-console': 0,
